requirements.txt.txt
fastapi
uvicorn
yfinance
requests
finnhub-python
